# Conflicts:
#	modules/matterhorn-scheduler-impl/pom.xml
#	modules/matterhorn-scheduler-impl/src/main/java/org/opencastproject/scheduler/endpoint/SchedulerRestService.java
#	modules/matterhorn-scheduler-remote/src/main/java/org/opencastproject/scheduler/remote/SchedulerServiceRemoteImpl.java

+++ modules/matterhorn-series-service-api/src/main/java/org/opencastproject/series/api/SeriesQuery.java
@@ -1,21 +1,27 @@
 /**
- *  Copyright 2009, 2010 The Regents of the University of California
- *  Licensed under the Educational Community License, Version 2.0
- *  (the "License"); you may not use this file except in compliance
- *  with the License. You may obtain a copy of the License at
+ * Licensed to The Apereo Foundation under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.
  *
- *  http://www.osedu.org/licenses/ECL-2.0
  *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an "AS IS"
- *  BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- *  or implied. See the License for the specific language governing
- *  permissions and limitations under the License.
+ * The Apereo Foundation licenses this file to you under the Educational
+ * Community License, Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of the License
+ * at:
+ *
+ *   http://opensource.org/licenses/ecl2.txt
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
  *
  */
+
 package org.opencastproject.series.api;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import java.util.Date;
 
@@ -26,9 +32,9 @@ import java.util.Date;
 public class SeriesQuery {
 
   /** Maximum number of results returned */
-  protected long count;
+  protected int count;
   /** Start page number */
-  protected long startPage;
+  protected int startPage;
   /** Free text search */
   protected String text;
   /** Series id search */
@@ -108,7 +114,7 @@ public class SeriesQuery {
    * @param count
    * @return
    */
-  public SeriesQuery setCount(long count) {
+  public SeriesQuery setCount(int count) {
     this.count = count;
     return this;
   }
@@ -119,7 +125,7 @@ public class SeriesQuery {
    * @param startPage
    * @return
    */
-  public SeriesQuery setStartPage(long startPage) {
+  public SeriesQuery setStartPage(int startPage) {
     this.startPage = startPage;
     return this;
   }
@@ -344,7 +350,7 @@ public class SeriesQuery {
    *
    * @return
    */
-  public long getCount() {
+  public int getCount() {
     return count;
   }
 
@@ -353,7 +359,7 @@ public class SeriesQuery {
    *
    * @return
    */
-  public long getStartPage() {
+  public int getStartPage() {
     return startPage;
   }
